+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/actions/EncryptStringAction.java
@@ -1,8 +1,9 @@
-package org.thermoweb.intellij.plugin.encrypt;
+package org.thermoweb.intellij.plugin.encrypt.actions;
 
 import java.util.Map;
 
 import org.jetbrains.annotations.NotNull;
+import org.thermoweb.intellij.plugin.encrypt.cipher.CipherUtils;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.command.WriteCommandAction;
@@ -20,9 +21,9 @@ public class EncryptStringAction extends JasyptAction {
 		Map<String, String> values = dialog.getValues();
 
 		String newValue = CipherUtils.encrypt(primaryCaret.getSelectedText(), values.get(PASSWORD_FIELD_NAME), values.get(ALGORITHM_FIELD_NAME));
-		final String cipheredPassword = "true".equals(values.get(ENCAPSULATE_FIELD_NAME)) ? "ENC(" + newValue + ")" : newValue;
+		final String cipheredString = "true".equals(values.get(ENCAPSULATE_FIELD_NAME)) ? "ENC(" + newValue + ")" : newValue;
 		WriteCommandAction.runWriteCommandAction(project,
-				() -> document.replaceString(primaryCaret.getSelectionStart(), primaryCaret.getSelectionEnd(), cipheredPassword));
+				() -> document.replaceString(primaryCaret.getSelectionStart(), primaryCaret.getSelectionEnd(), cipheredString));
 		updateSettings();
 		primaryCaret.removeSelection();
 	}