+++ src/test/java/org/thermoweb/intellij/plugin/encrypt/CipherUtilsTest.java
@@ -1,13 +1,15 @@
 package org.thermoweb.intellij.plugin.encrypt;
 
 import org.junit.Test;
+import org.thermoweb.intellij.plugin.encrypt.cipher.CipherUtils;
+import org.thermoweb.intellij.plugin.encrypt.exceptions.JasyptPluginException;
 
 import static org.junit.Assert.*;
 
 public class CipherUtilsTest {
 
 	@Test
-	public void should_encrypt_and_decrypt_string_correctly() {
+	public void should_encrypt_and_decrypt_string_correctly() throws JasyptPluginException {
 		String secretString = "secretString";
 		String password = "password";
 		String encryptedText = CipherUtils.encrypt(secretString, password, Algorithms.PBEWITHHMACSHA1ANDAES_128.getCode());