+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # encrypt-string-plugin Changelog
 
 ## [Unreleased]
+### Added
+- when an error occurs during decrypt, a notification is shown in the ide.
+- Line marker on yaml file with encrypted string (with ENC(...))
 
 ## [1.1.0] - 2023-07-05